fixed #3

+++ src/ca/usherbrooke/pacman/model/sound/ISoundPlayer.java
@@ -13,7 +13,7 @@ public interface ISoundPlayer {
   public void loop();
 
   public void stop();
-  
-  public void setVolume(float volume);
+
+  public void setVolume(int volume);
 
 }
